persist notification type on documents
useful for filtering and searching in mongo later
