+++ server/src/core/server/stacks/rejectComment.ts
@@ -8,14 +8,12 @@ import {
   hasTag,
   UpdateCommentStatus,
 } from "coral-server/models/comment";
+import { NotificationType } from "coral-server/models/notifications/notification";
 import { Tenant } from "coral-server/models/tenant";
 import { removeTag } from "coral-server/services/comments";
 import { moderate } from "coral-server/services/comments/moderation";
 import { I18n } from "coral-server/services/i18n";
-import {
-  InternalNotificationContext,
-  NotificationType,
-} from "coral-server/services/notifications/internal/context";
+import { InternalNotificationContext } from "coral-server/services/notifications/internal/context";
 import { AugmentedRedis } from "coral-server/services/redis";
 import { submitCommentAsSpam } from "coral-server/services/spam";
 import { Request } from "coral-server/types/express";
